Removed 'no-gitinfofile' option.
Using 'no-gitinfofile' will have negative effect on sites with
versioned dependencies. An example is the rest_server module which is
part of the services project:

  dependencies[] = libraries (>=2.x)

Since the 'no-gitinfofile' option will result in not writing the
version information to the libraries module the dependency will not be
resolved on site install.